REFACTOR :: SearchBox 리팩토링
