+++ apps/client/src/app/page.tsx
@@ -3,7 +3,7 @@
 import FilterTabs from "ui/components/FilterTabs/FilterTabs";
 import Products from "ui/components/Products/Products";
 import SummaryContent from "ui/components/SummaryContent/SummaryContent";
-import SearchBox from "ui/components/main/SearchBox/SearchBox";
+import SearchBox from "ui/components/SearchBox/SearchBox";
 import AppLayout from "client/layouts/AppLayout";
 import styled from "styled-components";
 
@@ -12,7 +12,7 @@ export default function Home() {
     <AppLayout>
       <Container>
         <SummaryContent subject="와라! 매점" description="와라! 매점을 통해 매점 이용을 편리하게!"/>
-        <SearchBox />
+        <SearchBox placeholder="원하는 상품을 검색해 보세요." buttonText="검색"/>
         <FilterTabs items={"상품"} />
         <Products/>
       </Container>
