+++ apps/client/src/app/qna/page.tsx
@@ -2,7 +2,7 @@
 
 import SummaryContent from "ui/components/SummaryContent/SummaryContent"
 import Questions from "ui/components/qna/Questions/Questions"
-import SearchBox from "ui/components/qna/SearchBox/SearchBox"
+import SearchBox from "ui/components/SearchBox/SearchBox";
 import AppLayout from "client/layouts/AppLayout"
 import styled from "styled-components"
 
@@ -12,7 +12,7 @@ export default function Qna(){
         <AppLayout>
             <Container>
                 <SummaryContent subject="Q&A" description="매점을 이용하면서 궁금했던/바라는 점을 물어볼 수 있어요."/>
-                <SearchBox />
+                <SearchBox placeholder="Q&A를 검색해 보세요." buttonText="질문 작성하기"/>
                 <Questions />
             </Container>
         </AppLayout>
